+++ tests/govtool-frontend/playwright/lib/helpers/cardano.ts
@@ -1,0 +1,5 @@
+export function lovelaceToAda(lovelace: number) {
+  if (lovelace === 0) return 0;
+
+  return lovelace / 1e6;
+}