+++ tests/govtool-frontend/playwright/tests/6-miscellaneous/miscellaneous.spec.ts
@@ -23,7 +23,7 @@ test("6C. Navigation within the dApp @smoke @fast", async ({
   ]);
 
   await expect(guidesPage).toHaveURL(
-    `${environments.docsUrl}/about/what-is-sanchonet-govtool`
+    `${environments.docsUrl}/about/what-is-sanchonet-govtool`,
   );
 
   if (isMobile(page)) {